Move LightclientServer code to module index
